[Refactor][Runner] Remove JAX directory as there is only one runner.
