+++ tests/runner/test_speculative_decoding_manager.py
@@ -8,11 +8,9 @@ from vllm.sampling_params import SamplingType
 from vllm.v1.outputs import DraftTokenIds
 from vllm.v1.spec_decode.ngram_proposer import NgramProposer
 
-from tpu_commons.runner.jax.input_batch_jax import (CachedRequestState,
-                                                    InputBatch)
-from tpu_commons.runner.jax.speculative_decoding_manager import \
-    SpecDecodeMetadata
-from tpu_commons.runner.jax.tpu_jax_runner import TPUModelRunner
+from tpu_commons.runner.input_batch_jax import CachedRequestState, InputBatch
+from tpu_commons.runner.speculative_decoding_manager import SpecDecodeMetadata
+from tpu_commons.runner.tpu_jax_runner import TPUModelRunner
 
 
 class TestSpeculativeDecodingManager:
@@ -26,7 +24,7 @@ class TestSpeculativeDecodingManager:
         with patch('jax.devices', return_value=self.mock_devices), \
              patch('jax.make_mesh', return_value=self.mock_mesh), \
              patch('jax.random.key', return_value=self.mock_rng_key), \
-             patch('tpu_commons.runner.jax.tpu_jax_runner.get_model', return_value=MagicMock()):
+             patch('tpu_commons.runner.tpu_jax_runner.get_model', return_value=MagicMock()):
 
             model_config = ModelConfig(tokenizer_mode="auto",
                                        trust_remote_code=False,
@@ -90,7 +88,7 @@ class TestSpeculativeDecodingManager:
         # Patch is_spec_decode_unsupported to control which requests are marked
         # as unsupported for speculative decoding.
         with patch(
-                'tpu_commons.runner.jax.input_batch_jax.is_spec_decode_unsupported'
+                'tpu_commons.runner.input_batch_jax.is_spec_decode_unsupported'
         ) as mock_is_unsupported:
             # We want req-2 to be unsupported. Let's use a simple condition.
             mock_is_unsupported.return_value = False
@@ -338,7 +336,7 @@ class TestSpeculativeDecodingManager:
 
         # Act
         with patch(
-                "tpu_commons.runner.jax.speculative_decoding_manager.device_array",
+                "tpu_commons.runner.speculative_decoding_manager.device_array",
                 side_effect=self.mock_device_array):
             metadata = self.runner.speculative_decoding_manager.get_spec_decode_metadata(
                 num_draft_tokens_np,
