+++ tests/runner/test_structured_decoding_manager.py
@@ -6,8 +6,8 @@ from vllm.config import (CacheConfig, ModelConfig, ParallelConfig,
                          SchedulerConfig, SpeculativeConfig, VllmConfig)
 from vllm.sampling_params import SamplingType
 
-from tpu_commons.runner.jax.input_batch_jax import CachedRequestState
-from tpu_commons.runner.jax.tpu_jax_runner import TPUModelRunner
+from tpu_commons.runner.input_batch_jax import CachedRequestState
+from tpu_commons.runner.tpu_jax_runner import TPUModelRunner
 
 
 class TestStructuredDecodingManager:
@@ -21,7 +21,7 @@ class TestStructuredDecodingManager:
         with patch('jax.devices', return_value=self.mock_devices), \
              patch('jax.make_mesh', return_value=self.mock_mesh), \
              patch('jax.random.key', return_value=self.mock_rng_key), \
-             patch('tpu_commons.runner.jax.tpu_jax_runner.get_model', return_value=MagicMock()):
+             patch('tpu_commons.runner.tpu_jax_runner.get_model', return_value=MagicMock()):
 
             model_config = ModelConfig(tokenizer_mode="auto",
                                        trust_remote_code=False,
@@ -75,7 +75,7 @@ class TestStructuredDecodingManager:
 
         # Patch the centralized device_array function instead of runner's method
         with patch(
-                'tpu_commons.runner.jax.structured_decoding_manager.device_array',
+                'tpu_commons.runner.structured_decoding_manager.device_array',
                 side_effect=mock_device_array):
 
             # Create a mock for sampling_params to avoid TypeErrors in add_request
