+++ tpu_commons/runner/speculative_decoding_manager.py
@@ -12,7 +12,7 @@ from tpu_commons.runner import utils as runner_utils
 from tpu_commons.utils import device_array
 
 if TYPE_CHECKING:
-    from tpu_commons.runner.jax.tpu_jax_runner import TPUModelRunner
+    from tpu_commons.runner.tpu_jax_runner import TPUModelRunner
 
 
 @dataclass
